Added code display in chat
